+++ gazebo/rendering/DepthCamera.cc
@@ -39,13 +39,8 @@ DepthCamera::DepthCamera(const std::string &_namePrefix, ScenePtr _scene,
   : Camera(_namePrefix, _scene, _autoRender),
     dataPtr(new DepthCameraPrivate)
 {
-  this->depthTarget = NULL;
-  this->dataPtr->depthBuffer = NULL;
-  this->dataPtr->depthMaterial = NULL;
-  this->dataPtr->pcdTarget = NULL;
-  this->dataPtr->pcdBuffer = NULL;
-  this->dataPtr->pcdMaterial = NULL;
   this->dataPtr->outputPoints = false;
+  this->dataPtr->outputNormals = false;
 }
 
 //////////////////////////////////////////////////
@@ -53,6 +48,9 @@ DepthCamera::~DepthCamera()
 {
   if (this->dataPtr->depthBuffer)
     delete [] this->dataPtr->depthBuffer;
+
+  if (this->dataPtr->normalsBuffer)
+    delete [] this->dataPtr->normalsBuffer;
 
   if (this->dataPtr->pcdBuffer)
     delete [] this->dataPtr->pcdBuffer;
@@ -62,9 +60,14 @@ DepthCamera::~DepthCamera()
 void DepthCamera::Load(sdf::ElementPtr _sdf)
 {
   Camera::Load(_sdf);
-  this->dataPtr->outputPoints =
-    (_sdf->GetElement("depth_camera")->Get<std::string>("output")
-    == "points");
+
+  std::string outputs = _sdf->GetElement("depth_camera")->
+                              Get<std::string>("output");
+
+  std::size_t found = outputs.find("points");
+  this->dataPtr->outputPoints =  found != std::string::npos;
+  found = outputs.find("normals");
+  this->dataPtr->outputNormals =  found != std::string::npos;
 }
 
 //////////////////////////////////////////////////
@@ -180,11 +183,52 @@ void DepthCamera::CreateDepthTexture(const std::string &_textureName)
 }
 
 //////////////////////////////////////////////////
+void DepthCamera::CreateNormalsTexture(const std::string &_textureName)
+{
+  if (this->dataPtr->outputNormals)
+  {
+    this->dataPtr->normalsTextures =
+      Ogre::TextureManager::getSingleton().createManual(
+      _textureName + "_normals",
+      "General",
+      Ogre::TEX_TYPE_2D,
+      this->ImageWidth(), this->ImageHeight(), 0,
+      Ogre::PF_FLOAT32_RGBA,
+      Ogre::TU_RENDERTARGET).getPointer();
+
+    this->dataPtr->normalsTarget =
+        this->dataPtr->normalsTextures->getBuffer()->getRenderTarget();
+    this->dataPtr->normalsTarget->setAutoUpdated(false);
+
+    this->dataPtr->normalsViewport =
+        this->dataPtr->normalsTarget->addViewport(this->camera);
+    this->dataPtr->normalsViewport->setClearEveryFrame(true);
+
+    auto const &ignBG = this->scene->BackgroundColor();
+    this->dataPtr->normalsViewport->setBackgroundColour(
+        Conversions::Convert(ignBG));
+    this->dataPtr->normalsViewport->setOverlaysEnabled(false);
+    this->dataPtr->normalsViewport->setVisibilityMask(
+        GZ_VISIBILITY_ALL & ~(GZ_VISIBILITY_GUI | GZ_VISIBILITY_SELECTABLE));
+
+    this->dataPtr->normalsMaterial = (Ogre::Material*)(
+    Ogre::MaterialManager::getSingleton().getByName("Gazebo/XYZNormals").get());
+
+    this->dataPtr->normalsMaterial->getTechnique(0)->getPass(0)->
+        createTextureUnitState(_textureName + "_normals");
+
+    this->dataPtr->normalsMaterial->load();
+  }
+}
+
+//////////////////////////////////////////////////
 void DepthCamera::PostRender()
 {
   this->depthTarget->swapBuffers();
   if (this->dataPtr->outputPoints)
     this->dataPtr->pcdTarget->swapBuffers();
+  if (this->dataPtr->outputNormals)
+    this->dataPtr->normalsTarget->swapBuffers();
 
   if (this->newData && this->captureData)
   {
@@ -239,6 +283,30 @@ void DepthCamera::PostRender()
       this->dataPtr->newRGBPointCloud(
           this->dataPtr->pcdBuffer, width, height, 1, "RGBPOINTS");
     }
+
+    if (this->dataPtr->outputNormals)
+    {
+      Ogre::HardwarePixelBufferSharedPtr normalsPixelBuffer;
+
+      normalsPixelBuffer = this->dataPtr->normalsTextures->getBuffer();
+
+      // Blit the depth buffer if needed
+      if (!this->dataPtr->normalsBuffer)
+        this->dataPtr->normalsBuffer = new float[width * height * 4];
+
+      memset(this->dataPtr->normalsBuffer, 0, width * height * 4);
+
+      Ogre::Box normals_src_box(0, 0, width, height);
+      Ogre::PixelBox normals_dst_box(width, height,
+          1, Ogre::PF_FLOAT32_RGBA, this->dataPtr->normalsBuffer);
+
+      normalsPixelBuffer->lock(Ogre::HardwarePixelBuffer::HBL_NORMAL);
+      normalsPixelBuffer->blitToMemory(normals_src_box, normals_dst_box);
+      normalsPixelBuffer->unlock();
+
+      this->dataPtr->newNormalsPointCloud(
+          this->dataPtr->normalsBuffer, width, height, 1, "NORMALS");
+    }
   }
 
   // also new image frame for camera texture
@@ -252,7 +320,7 @@ void DepthCamera::UpdateRenderTarget(Ogre::RenderTarget *_target,
           Ogre::Material *_material, const std::string &_matName)
 {
   Ogre::RenderSystem *renderSys;
-  Ogre::Viewport *vp = NULL;
+  Ogre::Viewport *vp = nullptr;
   Ogre::SceneManager *sceneMgr = this->scene->OgreSceneManager();
   Ogre::Pass *pass;
 
@@ -271,8 +339,11 @@ void DepthCamera::UpdateRenderTarget(Ogre::RenderTarget *_target,
   vp = _target->getViewport(0);
 
   // return farClip in case no renderable object is inside frustrum
-  vp->setBackgroundColour(Ogre::ColourValue(this->FarClip(),
-      this->FarClip(), this->FarClip()));
+  if (_target == this->dataPtr->normalsTarget)
+    vp->setBackgroundColour(Ogre::ColourValue(0, 0, 0));
+  else
+    vp->setBackgroundColour(Ogre::ColourValue(this->FarClip(),
+        this->FarClip(), this->FarClip()));
 
   Ogre::CompositorManager::getSingleton().setCompositorEnabled(
                                                 vp, _matName, true);
@@ -365,6 +436,19 @@ void DepthCamera::RenderImpl()
     sceneMgr->_suppressRenderStateChanges(false);
     sceneMgr->setShadowTechnique(shadowTech);
   }
+  if (this->dataPtr->outputNormals)
+  {
+    sceneMgr->setShadowTechnique(Ogre::SHADOWTYPE_NONE);
+    sceneMgr->_suppressRenderStateChanges(true);
+
+    this->UpdateRenderTarget(this->dataPtr->normalsTarget,
+                  this->dataPtr->normalsMaterial, "Gazebo/XYZNormals");
+
+    this->dataPtr->normalsTarget->update(false);
+
+    sceneMgr->_suppressRenderStateChanges(false);
+    sceneMgr->setShadowTechnique(shadowTech);
+  }
 }
 
 //////////////////////////////////////////////////
@@ -413,4 +497,12 @@ event::ConnectionPtr DepthCamera::ConnectNewRGBPointCloud(
     const std::string &)>  _subscriber)
 {
   return this->dataPtr->newRGBPointCloud.Connect(_subscriber);
+}
+
+//////////////////////////////////////////////////
+event::ConnectionPtr DepthCamera::ConnectNewNormalsPointCloud(
+    std::function<void (const float *, unsigned int, unsigned int, unsigned int,
+    const std::string &)>  _subscriber)
+{
+  return this->dataPtr->newNormalsPointCloud.Connect(_subscriber);
 }