+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ## Gazebo 9.XX.X (2020-XX-XX)
 
+1. Added normals to depth camera sensor
+    * [Pull request 3193](https://bitbucket.org/osrf/gazebo/pull-requests/3193/)
+
 1. Prevent crash when subscribing to depth camera image topic
     * [Pull request 3197](https://bitbucket.org/osrf/gazebo/pull-request/3197)